requirements.txt
customtkinter==5.1.1
Pillow==9.2.0
